src/main/webapp/app/entities/ordre-facturation/delete/ordre-facturation-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { OrdreFacturationService } from '../service/ordre-facturation.service';

import { OrdreFacturationDeleteDialogComponent } from './ordre-facturation-delete-dialog.component';

describe('OrdreFacturation Management Delete Component', () => {
  let comp: OrdreFacturationDeleteDialogComponent;
  let fixture: ComponentFixture<OrdreFacturationDeleteDialogComponent>;
  let service: OrdreFacturationService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [OrdreFacturationDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(OrdreFacturationDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(OrdreFacturationDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(OrdreFacturationService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
